TestDataAndOperation de News modularizado
